add mc-pathfinding library as dependency
