+++ .github/workflows/pull-request.yml
@@ -1,0 +1,45 @@
+name: Pull Request Validation
+
+on:
+  pull_request:
+
+jobs:
+  cleanup-and-build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up JDK 17.0.2
+        uses: actions/setup-java@v3
+        with:
+          distribution: 'temurin'
+          java-version: 17.0.2
+          server-id: github
+          cache: 'maven'
+
+      - name: Set PR Maven Version
+        id: set_pr_maven_version
+        run: echo "PR_MAVEN_VERSION=$(mvn help:evaluate -Dexpression=project.version -q -DforceStdout)" >> "$GITHUB_OUTPUT"
+
+      - name: Set Main Maven Version
+        id: set_main_maven_version
+        run: |
+          git fetch origin ${{ github.base_ref }}
+          git checkout ${{ github.base_ref }}
+          echo "MAIN_MAVEN_VERSION=$(mvn help:evaluate -Dexpression=project.version -q -DforceStdout)" >> "$GITHUB_OUTPUT"
+
+      - name: Compare Maven Versions
+        run: |
+          export pr_maven_version="${{ steps.set_pr_maven_version.outputs.PR_MAVEN_VERSION }}"
+          export main_maven_version="${{ steps.set_main_maven_version.outputs.MAIN_MAVEN_VERSION }}"
+          
+          if [ "$pr_maven_version" != "$main_maven_version" ]; then
+            echo "New pom.xml version detected ($main_maven_version -> $pr_maven_version)."
+          else
+            echo "Invalid pom.xml version - release already exists with this version."
+            exit 1
+          fi
+
+      - name: Build with Maven
+        if: success()
+        run: mvn clean package --file pom.xml
