+++ .gitignore
@@ -18,7 +18,7 @@ out/
 *.ctxt
 
 # Package Files #
-*.jar
+# .jar
 *.war
 *.nar
 *.ear
