+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>mc-lightning-strike</groupId>
     <artifactId>mc-lightning-strike</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <version>1.1.0</version>
     <packaging>jar</packaging>
 
     <name>mc-lightning-strike</name>
